Refactor to have tests into different modules
