+++ tests/test_base.py
@@ -1,0 +1,14 @@
+from importlib import import_module
+
+import pytest
+
+from tests.base import VENUES, TestsBase
+
+# from unittest.mock import patch
+
+
+class TestSDK(TestsBase):
+    @pytest.mark.parametrize("venue", VENUES)
+    def test_can_import_sdk_module(self, venue):
+        """equivalent to import venue"""
+        import_module(venue)
